requirements.txt
gymnasium
gym-super-mario-bros
matplotlib
nes_py
opencv-python
shimmy
stable_baselines3
torch
